# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/risk-responses": {
            "get": {
                "tags": [
                    "RiskResponses"
                ],
                "summary": "Lists risk-responses",
                "operationId": "listRiskResponses",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "count",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RiskResponsePagingResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "RiskResponses"
                ],
                "summary": "Adds a new risk-response",
                "operationId": "storeRiskResponse",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RiskResponseStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RiskResponseResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/risk-responses/{id}": {
            "get": {
                "tags": [
                    "RiskResponses"
                ],
                "summary": "Get RiskResponse information",
                "description": "Returns RiskResponse data",
                "operationId": "showRiskResponse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "RiskResponse id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RiskResponseResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "RiskResponses"
                ],
                "summary": "Update RiskResponse information",
                "description": "Returns updated RiskResponse data",
                "operationId": "updateRiskResponse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "RiskResponse id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RiskResponseStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RiskResponseResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "RiskResponses"
                ],
                "summary": "Delete RiskResponse",
                "description": "Deletes a single RiskResponse",
                "operationId": "deleteRiskResponse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "RiskResponse id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RiskResponseResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AssessmentLifecycleStatus": {
                "type": "string",
                "enum": [
                    "OPEN",
                    "CLOSED"
                ]
            },
            "RiskResponseLifecycleStatus": {
                "type": "string",
                "enum": [
                    "OPEN",
                    "IN_PROGRESS",
                    "CLOSED"
                ]
            },
            "Assessment": {
                "title": "Assessment",
                "description": "Assessment model",
                "type": "object",
                "xml": {
                    "name": "Assessment"
                },
                "allOf": [
                    {
                        "$ref": "#/components/schemas/NamedBaseModel"
                    },
                    {
                        "properties": {
                            "created": {
                                "title": "created",
                                "description": "Created date of the Assessment",
                                "type": "string",
                                "format": "date",
                                "example": "2020-01-27 17:50:45"
                            },
                            "lifecycle_status": {
                                "$ref": "#/components/schemas/AssessmentLifecycleStatus"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "BaseModel": {
                "title": "BaseModel",
                "description": "BaseModel",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BaseModel"
                }
            },
...
            "NamedBaseModel": {
                "title": "NamedBaseModel",
                "description": "NamedBaseModel",
                "type": "object",
                "xml": {
                    "name": "NamedBaseModel"
                },
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BaseModel"
                    },
                    {
                        "properties": {
                            "name": {
                                "title": "Name",
                                "description": "Name of the entity",
                                "type": "string",
                                "example": "Elmofuchto"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "RiskResponse": {
                "title": "RiskResponse",
                "description": "RiskResponse model",
                "type": "object",
                "xml": {
                    "name": "RiskResponse"
                },
                "allOf": [
                    {
                        "$ref": "#/components/schemas/NamedBaseModel"
                    },
                    {
                        "properties": {
                            "created": {
                                "title": "created",
                                "description": "Created date of the RiskResponse",
                                "type": "string",
                                "format": "date",
                                "example": "2020-01-27 17:50:45"
                            },
                            "lifecycle_status": {
                                "$ref": "#/components/schemas/RiskResponseLifecycleStatus"
                            },
                            "assessment_id": {
                                "title": "Assessment ID",
                                "description": "Assessment's id of the RiskResponse",
                                "type": "integer",
                                "format": "int64",
                                "example": 1
                            },
                            "assessment": {
                                "$ref": "#/components/schemas/Assessment"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "RiskResponseStoreRequest": {
                "title": "RiskResponseStoreRequest",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the new Assessment",
                        "type": "string",
                        "example": "Approve"
                    },
                    "created": {
                        "title": "created",
                        "description": "The date the RiskResponse was created",
                        "type": "string",
                        "format": "date",
                        "example": "2021-10-10"
                    },
                    "lifecycle_status": {
                        "$ref": "#/components/schemas/RiskResponseLifecycleStatus"
                    },
                    "assessment_id": {
                        "title": "assessment_id",
                        "description": "The assessment id of the risk response",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
            "Meta": {
                "title": "Meta",
                "properties": {
                    "current_page": {
                        "title": "Current Page",
                        "description": "Current Page of the data",
                        "type": "integer"
                    },
                    "from": {
                        "title": "From",
                        "description": "From",
                        "type": "integer"
                    },
                    "last_page": {
                        "title": "Last Page",
                        "description": "Last Page of the data",
                        "type": "integer"
                    },
                    "per_page": {
                        "title": "Per Page",
                        "description": "Per Page of the data",
                        "type": "integer"
                    },
                    "to": {
                        "title": "To",
                        "description": "To",
                        "type": "integer"
                    },
                    "total": {
                        "title": "Total",
                        "description": "Total",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "RiskResponsePagingResource": {
                "title": "RiskResponsePagingResource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RiskResponse"
                        }
                    },
                    "meta": {
                        "$ref": "#/components/schemas/Meta"
                    }
                },
                "type": "object"
            },
            "RiskResponseResource": {
                "title": "RiskResponseResource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RiskResponse"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format Bearer 4|oeXad4kChJT43wli90LOd1VbFhtuGuEdvxvEHMAtcb025185",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}